backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
chromadb
sentence-transformers==3.4.1
groq
pymupdf==1.25.3
pandas==2.2.3
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
